Package structure updated.
