+++ gold_digger/database/db_model.py
